more work on errorhandler
